Refactor the test of diagnostic tests
